Restructured components folder
