+++ homeworks/brisa-rojas/17-class-and-functional-components/src/components/add-button.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class AddButton extends React.Component {
+  render() {
+    return (
+      <div className="add-avatar-tile tile" onClick={this.props.onClick}>
+        +
+      </div>
+    );
+  }
+}
+
+export default AddButton;
