+++ homeworks/brisa-rojas/17-class-and-functional-components/src/components/refresh-button.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class RefreshAllButton extends React.Component {
+  render() {
+    return (
+      <div className="refresh-all-button" onClick={this.props.onClick}>
+        REFRESH ALL
+      </div>
+    );
+  }
+}
+
+export default RefreshAllButton;
